updated the 1001 and hours from income status
