+++ openfisca_canada/tests/benefitfinder/ei_workshare/ei_workshare.yaml
@@ -5,7 +5,7 @@
       input:
         has_some_income_1001_or_more_reduced_hours: true
       output:
-        is_ei_workshare_eligible: true 
+        ei_workshare__is_eligible: true 
 
     - name: "person is not eligible for EI Workshare"
       period: 2018-08
@@ -13,4 +13,4 @@
       input:
         has_some_income_1001_or_more_reduced_hours: false
       output:
-        is_ei_workshare_eligible: false
+        ei_workshare__is_eligible: false
